make removing whitespace from forms a mixin
